+++ R/PredictionDataClassif.R
@@ -8,7 +8,7 @@ check_prediction_data.PredictionDataClassif = function(pdata, train_task, ...) {
   assert_factor(pdata$truth, len = n, null.ok = TRUE)
   # unsupervised task
   if (is.null(pdata$truth)) {
-    lvls = fget(train_task$col_info, train_task$target_names, "levels", "id")
+    lvls = fget_key(train_task$col_info, train_task$target_names, "levels", "id")
     pdata$truth = if (length(pdata$row_ids)) factor(NA, lvls) else factor(levels = lvls)
   }
   lvls = levels(pdata$truth)